requirements.txt
certifi @ file:///croot/certifi_1671487769961/work/certifi
click @ file:///tmp/build/80754af9/click_1646056591685/work
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cytoolz @ file:///croot/cytoolz_1667465931118/work
dask==2021.10.0
Flask @ file:///croot/flask_1671217343254/work
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.25.0
fsspec @ file:///croot/fsspec_1670336592807/work
imagecodecs @ file:///opt/conda/conda-bld/imagecodecs_1664561985385/work
imageio @ file:///tmp/abs_cd920173-f360-47c5-97b0-bf4d1076d5d4dvic0oys/croots/recipe/imageio_1658785036907/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562407465/work
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib @ file:///croot/joblib_1666298844297/work
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
locket @ file:///opt/conda/conda-bld/locket_1652903118915/work
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib @ file:///croot/matplotlib-suite_1667356714455/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
munkres==1.1.4
networkx @ file:///tmp/build/80754af9/networkx_1633639043937/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
packaging @ file:///croot/packaging_1671697413597/work
partd @ file:///opt/conda/conda-bld/partd_1647245470509/work
Pillow==9.3.0
ply==3.11
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1648727874099/work
PyYAML @ file:///croot/pyyaml_1670514731622/work
scikit-image @ file:///croot/scikit-image_1669241743693/work
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1642601761909/work
scipy @ file:///opt/conda/conda-bld/scipy_1661390393401/work
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tifffile @ file:///tmp/build/80754af9/tifffile_1627275862826/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
toolz @ file:///croot/toolz_1667464077321/work
torch==1.10.1+cpu
torchaudio==0.10.1+cpu
torchvision==0.11.2+cpu
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
Werkzeug @ file:///croot/werkzeug_1671215996852/work
zipp @ file:///croot/zipp_1672387121353/work
